utils/extract.sh fix try to mv in '.' directory
issue:

{run} rstrnt-prefetch.sh -fetch.done.mark http://netqe-bj.usersys.redhat.com/share/mhou/mhou-kernel.tar.gz
{debug} processing 'http://netqe-bj.usersys.redhat.com/share/mhou/mhou-kernel.tar.gz'
{debug} downloading 'http://netqe-bj.usersys.redhat.com/share/mhou/mhou-kernel.tar.gz' to '/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou'
{INFO} url=http://netqe-bj.usersys.redhat.com/share/mhou/mhou-kernel.tar.gz
{INFO} run: curl -o /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz $url -f -L -k --continue-at -
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
   0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0 100 32.4M  100 32.4M    0     0   111M      0 --:--:-- --:--:-- --:--:--  111M
{debug} extracting to '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'
{run} tar -C '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou' -zxf '/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'  --warning=no-timestamp
{run} mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
mv: cannot move '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/.' to '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz': Device or resource busy

debug output:
{debug} extracting to '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'
+ extract.sh /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou mhou-kernel.tar.gz
+ P=extract.sh
++ getopt -o hp:l --long help --long path: --long list -a -n '$P' -- /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou mhou-kernel.tar.gz
+ _at=' -- '\''/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'\'' '\''/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou'\'' '\''mhou-kernel.tar.gz'\'''
+ eval set -- ' -- '\''/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'\'' '\''/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou'\'' '\''mhou-kernel.tar.gz'\'''
++ set -- -- /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou mhou-kernel.tar.gz
+ folders=()
+ true
+ case "$1" in
+ shift
+ break
+ [[ 3 = 0 ]]
+ compressedFile=/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
+ [[ -f /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz ]]
+ targetdir=/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou
+ [[ -d /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou ]]
+ mkdir -p /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou
+ topdir=mhou-kernel.tar.gz
+ [[ mhou-kernel.tar.gz = */* ]]
+ [[ mhou-kernel.tar.gz = .. ]]
++ file -b /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
+ filetype='gzip compressed data, from Unix, original size modulo 2^32 211322880'
+ _targetdir=/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou
+ [[ gzip compressed data, from Unix, original size modulo 2^32 211322880 = Zip* ]]
+ case "$filetype" in
+ xtype=z
+ otopdir=($(tar taf ${compressedFile} | awk -F/ '{a[$1]++} END { for(key in a) { print(key) } }'))
++ tar taf /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
++ awk -F/ '{a[$1]++} END { for(key in a) { print(key) } }'
+ [[ -z . ]]
+ [[ 1 -gt 1 ]]
+ (( i=0 ))
+ (( i<0 ))
+ [[ '' = yes ]]
+ man tar
+ grep -q '^ *timestamp'
+ warnOpt=--warning=no-timestamp
+ [[ -n '' ]]
+ echo '{run} tar -C '\''/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou'\'' -zxf '\''/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'\''  --warning=no-timestamp'
{run} tar -C '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou' -zxf '/var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'  --warning=no-timestamp
+ tar -C /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou -zxf /var/cache/nrestraint/fetch-uri/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz --warning=no-timestamp
+ [[ -d /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. ]]
+ [[ -n mhou-kernel.tar.gz ]]
+ [[ mhou-kernel.tar.gz != \. ]]
+ test -d /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
+ [[ -e /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz ]]
+ _cmd='mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'
+ echo '{run} mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz'
{run} mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
+ eval mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
++ mv -T /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/. /mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz
mv: cannot move '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/.' to '/mnt/tests/netqe-bj.usersys.redhat.com//share/mhou/mhou-kernel.tar.gz': Device or resource busy

+++ utils/extract.sh
@@ -78,11 +78,29 @@ else
 	esac
 	#dirlist=$(tar taf ${compressedFile}|grep /$)
 	otopdir=($(tar taf ${compressedFile} | awk -F/ '{a[$1]++} END { for(key in a) { print(key) } }'))
-	[[ -z "$otopdir" ]] && { echo "{error} extract $compressedFile fail" >&2; exit 3; }
-	if [[ "${#otopdir[@]}" -gt 1 ]]; then
+	# Handle the case where otopdir is . or empty
+	if [[ "${#otopdir[@]}" -eq 0 || "${otopdir[0]}" == "." ]]; then
+		# If there is no explicit top-level directory, use the specified topdir or generate one.
+		if [[ -n "$topdir" ]]; then
+			# Use the specified topdir
+			_targetdir="$targetdir/$topdir"
+			mkdir -p "$_targetdir"
+			# Set otopdir to empty to avoid subsequent renaming operations
+			otopdir=("")
+		else
+			# enerate directory names based on file names
+			local default_dir="${compressedFile##*/}" # Get the full file name
+			_targetdir="$targetdir/$default_dir" # Build the target directory
+			mkdir -p "$_targetdir" # Create a directory
+			otopdir=("") # Clear the original directory name
+			topdir="$default_dir" # Set the new directory name
+		fi
+	elif [[ "${#otopdir[@]}" -gt 1 ]]; then
+		# Multiple top-level directories
 		otopdir=; [[ -n "$topdir" ]] && _targetdir+=/$topdir; topdir=
 		[[ "$_targetdir" != $targetdir ]] && { mkdir -p $_targetdir; }
 	else
+		# The case of a single top-level directory
 		for ((i=0; i<${#folders[@]}; i++)); do folders[$i]=$otopdir/${folders[$i]}; done
 	fi
 	[[ "$list" = yes ]] && { tar taf ${compressedFile} "${folders[@]}"; exit; }
@@ -102,11 +120,15 @@ fi
 	echo "{Error} extract to '$_targetdir' fail, please theck permission" >&2
 	exit 1
 }
-[[ -n "$topdir" && "${topdir}" != "$otopdir" ]] && {
+[[ -n "$topdir" && -n "${otopdir[0]}" && "$topdir" != "${otopdir[0]}" ]] && {
 	if test -d "$_targetdir/$topdir"; then
-		_cmd="(cd '$_targetdir/$otopdir' && tar c .) | (cd '$_targetdir/$topdir' && tar xf -); rm -rf '$_targetdir/$otopdir'"
+		_cmd="(cd '$_targetdir/${otopdir[0]}' && tar c .) | (cd '$_targetdir/$topdir' && tar xf -); rm -rf '$_targetdir/${otopdir[0]}'"
 	else
-		_cmd="mv -T $_targetdir/${otopdir} $_targetdir/$topdir"
+		if [[ -e "$_targetdir/$topdir" ]]; then
+			echo "{warn} target '$topdir' already exists, skipping rename" >&2
+		else
+			_cmd="mv -T $_targetdir/${otopdir} $_targetdir/$topdir"
+		fi
 	fi
 	echo "{run} $_cmd" >&2
 	eval $_cmd
